requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.7
filelock==3.12.4
flask==3.0.0
fsspec==2023.12.2
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
platformdirs==3.10.0
python-dotenv==1.0.0
requests==2.31.0
sympy==1.12
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
typing-extensions==4.9.0
urllib3==2.1.0
virtualenv==20.24.5
werkzeug==3.0.1
zipp==3.17.0
